+++ packages/ui-components/src/components/Chart/Table/ChartTable.js
@@ -6,7 +6,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import MuiTableContainer from '@material-ui/core/TableContainer';
-import { DataTable } from '../../DataTable';
+import { DataPreviewTable } from '../../DataPreviewTable';
 import { getIsChartData, getNoDataString } from '../utils';
 import { SmallAlert } from '../../Alert';
 import { getChartTableData } from './getChartTableData';
@@ -34,7 +34,7 @@ export const ChartTable = ({ viewContent, className }) => {
 
   return (
     <TableContainer className={className}>
-      <DataTable columns={columns} data={data} />
+      <DataPreviewTable columns={columns} data={data} />
     </TableContainer>
   );
 };
