+++ packages/ui-components/src/components/DataPreviewTable/index.js
@@ -3,5 +3,5 @@
  *  Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
  */
 
-export * from './DataTable';
+export * from './DataPreviewTable';
 export * from './useDataTableExport';
